decouple Wallet from WalletManager. Create shelley Wallet and other related placeholder classes
force git to detect file name changes
