+++ src/components/Delegation/DelegationCenter.js
@@ -10,7 +10,7 @@ import {STAKING_CENTER_ROUTES} from '../../RoutesList'
 import {withNavigationTitle} from '../../utils/renderUtils'
 import {CONFIG} from '../../config/config'
 import {Logger} from '../../utils/logging'
-import walletManager from '../../crypto/wallet'
+import walletManager from '../../crypto/walletManager'
 import {getShelleyTxFee} from '../../crypto/jormungandr/transactions/utils'
 import {InsufficientFunds} from '../../crypto/errors'
 import globalMessages, {errorMessages} from '../../i18n/global-messages'
