+++ src/components/Delegation/DelegationSummary.js
@@ -47,7 +47,7 @@ import {fetchUTXOs} from '../../actions/utxo'
 import {fetchPoolInfo} from '../../actions/pools'
 import {checkForFlawedWallets} from '../../actions'
 import {JORMUN_WALLET_ROUTES, WALLET_INIT_ROUTES} from '../../RoutesList'
-import walletManager from '../../crypto/wallet'
+import walletManager from '../../crypto/walletManager'
 import globalMessages from '../../i18n/global-messages'
 import {formatAdaWithText, formatAdaInteger} from '../../utils/format'
 import FlawedWalletScreen from './FlawedWalletScreen'
