+++ src/components/Send/SendScreen.js
@@ -32,7 +32,7 @@ import {fetchUTXOs} from '../../actions/utxo'
 import {withNavigationTitle} from '../../utils/renderUtils'
 import {formatAdaWithText, formatAdaWithSymbol} from '../../utils/format'
 import {parseAdaDecimal} from '../../utils/parsing'
-import walletManager from '../../crypto/wallet'
+import walletManager from '../../crypto/walletManager'
 import {validateAmount, validateAddressAsync} from '../../utils/validators'
 import AmountField from './AmountField'
 import UtxoAutoRefresher from './UtxoAutoRefresher'
